Cascoda SDK patch release v0.21-4
Add guide for setting up a network with two border routers
Update chilictl documentation
Update documentation URLs to IoTivity website
Download new PKI test certs with every build - always up to date
Add several unit tests for persistent storage API

+++ baremetal/platform/cascoda-nuvoton-chili2/port/source/cascoda_bsp_w25q80dlsnig.c
@@ -155,8 +155,7 @@ typedef enum statusRegister1_pos
 */
 static bool pageOverflow(uint32_t startAddress, uint8_t numOfBytes)
 {
-	int numOfBytesLeftInThePage =
-	    NUM_OF_BYTES_IN_A_PAGE - (int)(startAddress % NUM_OF_BYTES_IN_A_PAGE) - (int)numOfBytes;
+	int numOfBytesLeftInThePage = NUM_OF_BYTES_IN_A_PAGE - (startAddress % NUM_OF_BYTES_IN_A_PAGE) - numOfBytes;
 
 	if (numOfBytesLeftInThePage < 0)
 		return true;
